:arrow_up: Upgrade ava, gitmoji-cli, and git-url-parse (#45)
Running `npm audit` revealed several vulnerabilities, which I fixed using `npm audit fix --force`. The ava, gitmoji-cli, and git-url-parse packages had a major version upgrade. The ava and gitmoji-cli packages are dev dependencies. I ensured that both of them are working as expected. The git-url-parse package is a normal dependency. It's used to get the owner, repository name, and remote of the git repository. I ensured that the tests are passing but I didn't manually verify that the code is working as expected.